alteracoes finais no front
